Create check_pf_mem
Although it is written in Perl, this is a far cleaner version than my check_pf_mem hackjob. The original version of this plugin may be Nagios Exchange. I modified the location of the perl intepreter and I also modified the script to use the "used" parameter by default if neither option was provided.